requirements.txt
mongoose
pymongo==4.7.0
dnspython==2.4.2
flask
python-dotenv
flask_cors
dotenv
react-simple-maps
d3-geo
react-DOM
requests 
pymongo
python-dateutil
